pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "preview-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "pyyaml>=6.0",
]

[project.scripts]
preview-agent = "preview_agent.main:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
